refactor: move interfaces and types in seporated folders
