+++ src/components/Button/interfaces.ts
@@ -1,0 +1,7 @@
+export interface ButtonProps {
+  value: string
+  buttonType: string
+  width?: string
+  height?: string
+  handleExpressionValue: (value: string) => void
+}
