+++ src/components/CalculatorWrapper/interfaces.ts
@@ -1,0 +1,8 @@
+export interface CalculatorWrapperState {
+  expression: string
+  result: string
+  currentOperator: string
+  history: Array<string>
+  isError: boolean
+  isFinished: boolean
+}
