+++ src/components/Display/interfaces.ts
@@ -1,0 +1,5 @@
+export interface DisplayProps {
+  value: string
+  result: string
+  error: boolean
+}
