+++ src/components/ErrorBoundary/interfaces.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react'
+
+export interface ErrorBoundaryState {
+  error: boolean
+}
+
+export interface ErrorBoundaryProps {
+  children: ReactNode
+}
