+++ src/components/History/interfaces.ts
@@ -1,0 +1,3 @@
+export interface HistoryProps {
+  historyData: Array<string>
+}
