+++ src/components/Keyboard/interfaces.ts
@@ -1,0 +1,3 @@
+export interface KeyboardProps {
+  handleButton: (value: string) => void
+}
