+++ src/components/Keyboard/styles.ts
