+++ src/components/Layout/index.tsx
@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom'
 
 import { Header } from '@/components/Header'
 
-import { StyledLayout, StyledContainer } from './components'
+import { StyledLayout, StyledContainer } from './styles'
 
 export const Layout = () => (
   <StyledLayout>
