+++ src/constants/calculation.ts
@@ -1,0 +1,17 @@
+export enum SecondaryOperators {
+  CLEAR_ALL = 'AC',
+  EQUAL = '=',
+  COMMA = '.',
+  CLEAR = '->',
+  OPPOSITE_SIGN = '+/-',
+  OPEN_BRACKET = '(',
+  CLOSE_BRACKET = ')',
+}
+
+export enum MainOperators {
+  PLUS = '+',
+  MINUS = '-',
+  DIVIDE = '/',
+  MULTIPLY = 'x',
+  REMAINDER = '%',
+}
