+++ src/constants/operators.ts
@@ -1,0 +1,3 @@
+export const numbers = ['7', '8', '9', '4', '5', '6', '1', '2', '3', '0', '00', '.']
+export const rightOperators = ['/', '%', 'x', '-', '+', '=']
+export const leftOperators = ['(', ')', '->', '+/-', 'AC']
