+++ src/functionComponents/Button/types.ts
@@ -1,0 +1,7 @@
+export type ButtonProps = {
+  value: string
+  buttonType: string
+  width?: string
+  height?: string
+  handleExpressionValue: (value: string) => void
+}
