+++ src/functionComponents/Display/types.ts
@@ -1,0 +1,5 @@
+export type DisplayProps = {
+  value: string
+  result: string
+  error: boolean
+}
