+++ src/functionComponents/History/types.ts
@@ -1,0 +1,3 @@
+export type HistoryProps = {
+  historyData: Array<string>
+}
