+++ src/functionComponents/Keyboard/types.ts
@@ -1,0 +1,3 @@
+export type KeyboardProps = {
+  handleButton: (value: string) => void
+}
